Refactor: Image component에서 useStaticImageProps hooks분리
